module restructure party
