+++ src/map/Map.ts
@@ -6,6 +6,7 @@ import esri = __esri;
 import { EsriBundle, InfoBundle } from '../gapiTypes';
 import MapBase from './MapBase';
 import LayerBase from '../layer/BaseLayer';
+import HighlightLayer from '../layer/HighlightLayer';
 
 export default class Map extends MapBase {
 
@@ -30,6 +31,10 @@ export default class Map extends MapBase {
     // TODO implement
     addLayer (layer: LayerBase): void {
         this.innerMap.add(layer.innerLayer);
+    }
+
+    addHighlightLayer (highlightLayer: HighlightLayer) : void {
+        this.innerMap.add(highlightLayer.innerLayer);
     }
 
     // TODO passthrough functions, either by aly magic or make them hardcoded
