+++ src/old_trash/readthis.md
@@ -1,0 +1,3 @@
+Stuff here, we're not quite sure if we're gonna keep it.
+
+Once things seem done thanks, we should nuke this folder.
